refactor: JanggiGame 객체 분리 및 메서드 구체화
